refactor: move time utils
