+++ app/src/tests/components/contents/runner/run/PacketRunForm.test.tsx
@@ -1,12 +1,11 @@
 import { render, screen, waitFor } from "@testing-library/react";
 import { PacketRunForm } from "../../../../../app/components/contents/runner/run/PacketRunForm";
 import { mockGitBranches } from "../../../../mocks";
-// eslint-disable-next-line max-len
-import { getTimeDifferenceToDisplay } from "../../../../../app/components/contents/explorer/utils/getTimeDifferenceToDisplay";
 import userEvent from "@testing-library/user-event";
 import { server } from "../../../../../msw/server";
 import { rest } from "msw";
 import { MemoryRouter, Route, Routes } from "react-router-dom";
+import { getTimeDifferenceToDisplay } from "../../../../../lib/time";
 
 const renderComponent = () => {
   const mutate = jest.fn();
